Use a record instead of super long function sigs
I'm starting to lose track of what variables are what as I pass
everything around so I wrapped most of them in a record. The record
represents a single state in the recursive process of calculating where
mines could possibly be: mines we've placed (picked), numbers that are
maxed out (maxed) where we've picked the maximum number of mines that
can be touching and the sections of the grid that have cells touching a
certain number of mines (groups).