Solve for every possible solution
Given a specification for overlapping sets of numbered tiles that must
contain a certain number of mines calculate all possible solutions
containing the numbered cells containing mines.

An example spec would be:
[{[1,2,3,15,16],2},
 {[2,3,4],1},
 {[3,4,5,6,7],2},
 {[6,7,8],1},
 {[7,8,9,10,11],2},
 {[10,11,12],1},
 {[11,12,13,14,15],1},
 {[14,15,16],1}]

Cells 1-3, 15 and 16 must have 2 mines. That is, some cell connected to
all of those cells has reported that it is adjacent to two mines.

The example spec would be for a section of a "mines" / "minesweeper"
like the following:

1   2  3  4  5  The numbers on the outside are the cells that mines
16  2  1  2  6  can be in and the numbers on the inside are the
15  1     1  7  reported numbers of adjacent mines.
14  1  1  2  8
13  12 11 10 9