Cell probabilities based on all possible solutions
Give a list of possible solutions calculate the probability that a mine
will be in a particular cell. Each cell can only have one mine so it
will only appear once in a solution. If we take the number of times that
cell appears (i.e. it has a mine) divided by the number of solutions,
we'll have the probability that that actual solution will be one that
has a mine in that cell.